Created functionality to identify what word was replaced to the user
created functionality to identify which words were replaced via css (like changing the color or italicizing the word) to the user based on severity of the bad words.

Updated style.css to display different styles for the severity of the bad word.

Updated index to include some JINJA that uses a function in our app to find the severity of each word and display them and refer to style.css when displaying them.